requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
annotated-types=0.6.0=pypi_0
anyio=4.3.0=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.2.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
beautifulsoup4=4.12.3=pypi_0
bleach=6.1.0=pypi_0
blinker=1.7.0=pypi_0
ca-certificates=2023.11.17=h56e8100_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
coloredlogs=15.0.1=pypi_0
comm=0.2.1=pyhd8ed1ab_0
dash=2.15.0=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
debugpy=1.6.7=py39hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pyhd8ed1ab_0
fastapi=0.110.1=pypi_0
fastjsonschema=2.19.1=pypi_0
filelock=3.9.0=pypi_0
flask=3.0.2=pypi_0
flatbuffers=24.3.7=pypi_0
fsspec=2023.4.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.42=pypi_0
humanfriendly=10.0=pypi_0
idna=3.6=pypi_0
importlib-metadata=7.0.1=pyha770c72_0
importlib_metadata=7.0.1=hd8ed1ab_0
ipykernel=6.29.0=pyha63f2e9_0
ipython=8.12.3=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=4.12.0=py39hcbf5309_0
jupyterlab-pygments=0.3.0=pypi_0
libsodium=1.0.18=h8d14728_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=3.0.2=pypi_0
mpmath=1.3.0=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.3=pypi_0
nbformat=5.9.2=pypi_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
numpy=1.26.3=pypi_0
onnx=1.15.0=pypi_0
onnxruntime=1.17.1=pypi_0
onnxscript=0.1.0.dev20240312=pypi_0
openssl=3.0.12=h2bbff1b_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.2.0=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pip=23.3.1=py39haa95532_0
pipreqs=0.5.0=pypi_0
plotly=5.18.0=pypi_0
pprintpp=0.4.0=pypi_0
prompt-toolkit=3.0.42=pyha770c72_0
prompt_toolkit=3.0.42=hd8ed1ab_0
protobuf=4.25.3=pypi_0
psutil=5.9.0=py39h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pydantic=2.6.4=pypi_0
pydantic-core=2.16.3=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pyreadline3=3.4.1=pypi_0
python=3.9.18=h1aa4202_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2023.4=pypi_0
pywin32=227=py39hb82d6ee_1
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=py39hd77b12b_0
referencing=0.33.0=pypi_0
requests=2.31.0=pypi_0
retrying=1.3.4=pypi_0
rpds-py=0.17.1=pypi_0
sentry-sdk=1.42.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py39haa95532_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.37.2=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
tinycss2=1.2.1=pypi_0
torch=2.2.1+cu121=pypi_0
torchaudio=2.2.1+cu121=pypi_0
torchvision=0.17.1+cu121=pypi_0
tornado=6.2=py39hb82d6ee_0
traitlets=5.14.1=pyhd8ed1ab_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023.4=pypi_0
urllib3=2.2.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wandb=0.16.4=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
webencodings=0.5.1=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py39haa95532_0
yarg=0.1.9=pypi_0
zeromq=4.3.5=hd77b12b_0
zipp=3.17.0=pyhd8ed1ab_0
